lab5_USBGuard.adoc: fixup the end of the lab
The end of the lab ends with the USB drive blocked, which is fine, as
long as you don't do the "OPTIONAL" steps at the very end of the lab.
The optional steps help show that not *all* USB drives are blocked,
only the specified devices; by enabling the original USB drive, and
blocking the second, optional drive this is more easily demonstrated.