Fix issue #156
Addresses minor problems pointed out in #156.
The problem that opening shared.xml in nautilus is NOT simple (Text
Editor is not one of the options, it has to be selected from "other
applications") and Firefox, which is default, will fail to open it
is solved by listing the contents in the terminal instead, as we
work with terminal in the whole lab, so there is no need to use
Nautilus just in this specific case.